ajout de controllers

+++ src/main/java/ma/pfa/gestion_evenement/Controller/CompteController.java
@@ -26,7 +26,7 @@ public List<Compte> findAll() {
         }
 
 @DeleteMapping("/deleteAccount/{id}")
-public void delete(@PathVariable("id")  Integer id){
+public void delete(@PathVariable("id") Long id){
         compteRepository.deleteById(id);
 
         }
